+++ changelog.md
@@ -25,3 +25,11 @@
 
 ## [1.2.1] - 2025-07-29
 - Changement de fabric en internal_fabric pour éviter la confusion avec le module fabric SSH
+
+## [1.3.0] - 2025-08-01
+- Changement du vault:
+	- Ajout de la commande --unset
+	- Ajout de la commande --identity
+	- Ajout de la commande --keyring
+	- Ajout de la commande --tree
+	- Amélioration de la gestion des clés